Updated perf tests
