+++ test/PerformanceProfiling/FullTextIndexTests.cs
@@ -3,6 +3,7 @@ using BenchmarkDotNet.Jobs;
 using Lifti;
 using Lifti.Serialization.Binary;
 using Lifti.Tokenization;
+using Lifti.Tokenization.TextExtraction;
 using System;
 using System.IO;
 using System.Linq;
@@ -10,73 +11,6 @@ using System.Threading.Tasks;
 
 namespace PerformanceProfiling
 {
-    //[RankColumn, MemoryDiagnoser]
-    //[ShortRunJob(RuntimeMoniker.NetCoreApp31)]
-    //[ShortRunJob(RuntimeMoniker.Net60)]
-    //public class IndexSearchingBenchmarks : IndexBenchmarkBase
-    //{
-    //    private IFullTextIndex<string> index;
-
-    //    [GlobalSetup]
-    //    public async Task SetUp()
-    //    {
-    //        this.index = CreateNewIndex(4);
-    //        await this.PopulateIndexAsync(this.index);
-    //    }
-
-    //    [Params("(confiscation & th*) | \"and they\"")]
-    //    public string SearchCriteria { get; set; }
-
-    //    [Benchmark]
-    //    public object Searching()
-    //    {
-    //        return this.index.Search(this.SearchCriteria);
-    //    }
-    //}
-
-    //[SimpleJob(RuntimeMoniker.NetCoreApp22)]
-    //[SimpleJob(RuntimeMoniker.NetCoreApp31)]
-    //[RankColumn, MemoryDiagnoser]
-    //public class WordSplittingBenchmarks : IndexBenchmarkBase
-    //{
-    //    [Benchmark()]
-    //    public void XmlWorkSplittingNew()
-    //    {
-    //        var splitter = new XmlTokenizer();
-
-    //        splitter.Process(WikipediaData.SampleData[0].text).ToList();
-    //    }
-    //}
-
-    //[MediumRunJob(RuntimeMoniker.NetCoreApp31)]
-    [MediumRunJob(RuntimeMoniker.Net60)]
-    [RankColumn, MemoryDiagnoser]
-    public class SerializationBenchmarks : IndexBenchmarkBase
-    {
-        private BinarySerializer<string> serializer;
-        private string fileName;
-
-        [GlobalSetup]
-        public async Task Setup()
-        {
-            var index = CreateNewIndex(2);
-            await this.PopulateIndexAsync(index);
-
-            this.serializer = new BinarySerializer<string>();
-            this.fileName = $"{Guid.NewGuid()}.dat";
-            using var stream = File.OpenWrite(this.fileName);
-            await this.serializer.SerializeAsync(index, stream, true);
-        }
-
-        [Benchmark()]
-        public async Task IndexDeserialization()
-        {
-            var index = CreateNewIndex(2);
-            using var stream = File.OpenRead(this.fileName);
-            await this.serializer.DeserializeAsync(index, stream, true);
-        }
-    }
-
     [MediumRunJob(RuntimeMoniker.Net481)]
     [MediumRunJob(RuntimeMoniker.Net70)]
     [MediumRunJob(RuntimeMoniker.Net60)]
